added upgraded manga viewer
used electron to  build manga viewer as a windows app. auto update is supported. i'm pushing the source code in the repo. the exe will be released is releases

also added a lot of extra design. but yeah it made the app really heavy than before. it's best to use the html viewer unless you want to admire some stars. oops spoilers :)